# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sampooran Holidays Travel Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List holiday packages
      parameters:
        - name: destinationSlug
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: minDays
          in: query
          schema:
            type: integer
        - name: maxDays
          in: query
          schema:
            type: integer
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: type
          in: query
          schema:
            type: string
            enum: [b2b, b2c, both]
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: object
                properties:
                  packages:
                    type: array
                    items:
                      $ref: "#/components/schemas/Package"
                  total:
                    type: integer
                required:
                  - packages
                  - total
...
  /packages/{slug}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get package detail by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Package details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackageDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        destinationId:
          type: integer
        destinationName:
          type: string
        stateName:
          type: string
        countryName:
          type: string
        imageUrl:
          type: string
        thumbnailUrl:
          type: string
        shortDescription:
          type: string
        duration:
          type: integer
        nights:
          type: integer
        pricePerPerson:
          type: number
        originalPrice:
          type: number
        discountPercent:
          type: integer
        category:
          type: string
        packageType:
          type: string
          enum: [b2b, b2c, both]
        isFeatured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
        highlights:
          type: array
          items:
            type: string
        cities:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        inclusionIcons:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - slug
        - destinationId
        - destinationName
        - stateName
        - countryName
        - duration
        - nights
        - pricePerPerson
        - category
        - packageType
        - isFeatured
    ItineraryDay:
      type: object
      properties:
        day:
          type: integer
        title:
          type: string
        description:
          type: string
        activities:
          type: array
          items:
            type: string
        meals:
          type: array
          items:
            type: string
        accommodation:
          type: string
      required:
        - day
        - title
        - description
    PackageDetail:
      allOf:
        - $ref: "#/components/schemas/Package"
        - type: object
          properties:
            longDescription:
              type: string
            itinerary:
              type: array
              items:
                $ref: "#/components/schemas/ItineraryDay"
            inclusions:
              type: array
              items:
                type: string
            exclusions:
              type: array
              items:
                type: string
            hotels:
              type: array
              items:
                $ref: "#/components/schemas/HotelOption"
            tourType:
              type: string
            groupSize:
              type: string
            pickupPoint:
              type: string
            importantNotes:
              type: array
              items:
                type: string
            cancellationPolicy:
              type: string
            faqs:
              type: array
              items:
                $ref: "#/components/schemas/FAQ"
            relatedPackages:
              type: array
              items:
                $ref: "#/components/schemas/Package"
          required:
            - itinerary
            - inclusions
            - exclusions
    HotelOption:
      type: object
      properties:
        city:
          type: string
        hotelName:
          type: string
        category:
          type: string
        nights:
          type: integer
        roomType:
          type: string
      required:
        - city
        - hotelName
        - category
        - nights
    FAQ:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
      required:
        - question
        - answer
...